Demo5 with CSS transition
Will serve as an example for the migration guide from CSSTG to
react-motion. This demo is a sweet spot for TransitionSpring because
unmouting's done immediately. If a real `willLeave` is needed then we
either need to have side-effect in `willLeave` (yuck) to record the
current time, or have `willLeave` pass the elapsed time to us. I really
don't wanna pass time, even though it's trivial. We're doing alll of
this to avoid hardcoding timed animation. But maybe support for CSS
transition is by itself a compromise. If that makes sense, then it might
be better to keep core clean and split the whole thing into a
CSSTransitionSpring.