Make `willLeave` stop when `currVelocity` is 0 and `currValue` reached dest.
This makes `willLeave` simpler and have a better default behavior. The way to
tell the `TransitionSpring` before this was to add an if statement at the top of
`willLeave` to check if `currValue` has reached the destination and that the
`currVelocity` is 0 and then returning null if yes. We now do that check
internally and do the unmounting when `currValue` is equal to the return of
`endValue` and `currVelocity` is 0.

This "fixes" #104 by just deprecating the API.