Add Travis CI configuration
This will tell Travis CI how to test this project. This configuration
needs to be coupled with someone who has permissions on the GitHub repo
to enable the Travis CI webhook.

The nyan cat reporter didn't work very well in the Travis output, so I
decided to just use the default reporter for this integration.

[Closes #182]