Add support for interpolation precision; change spring() format
This changes the format of `spring(val, [stiffness, damping])` to
`spring(val, {stiffness, damping, onRest, precision})`.

Docs of this breaking change coming soon, but tldr: second argument now
an object (still optional). `{stiffness: 120, damping: 12}` is clearer
than the previous `[120, 12]`. onRest is for the future end callback
(name subject to change, no implementation yet). Precision is the
rounding of interpolating value. All alltributes are optionals and have
sensible defaults.

More on precision: #100. We don't actually round the numbers; just a
check on whether the destination delta and speed is smaller than a
certain threshold. Rounding the number during interpolation _might not_
be very useful. The main concern is that we can optimize by not
rendering, when the number's so small that it doesn't change. But I'm
not sure the probability of every (rounded) value in `style` not
changing, at any frame, compared to the last, is high enough to justify
the overhead of checking it every frame. We'll see.