(Continued) TransitionMotion Array styles
Part 2. See previous part for changes to the mergeDiff algorithm.
A big diff, but most of it are straightfoward logic of changing the
format from `{key1: {x: spring(0)}}` to `[{key: 1, style: {x:
spring(0)}}]`.

For the justification of the change, see
https://github.com/chenglou/react-motion/issues/26#issuecomment-165457623

This makes `styles` and the rest somewhat more verbose, but much more
straightforward to work with.

Documentations coming soon.