(Continued) TransitionMotion demo change
Part 4. See previous commits.

Just some straightforward changes to the demo to make it use the new
API. Array manipulation is indeed simpler for this use-case.

One thing to note is the subtlety of the change to
`willEnter`/`willLeave`. They don't take as many params anymore and ask
you to return the _same_ style object, rather than the style along with
its wrapper that constitutes the array cell, e.g. `{key: 1, style:
{...}}`. This will be documented better later ofc.